Implement `ViewInvoiceCommonService`
We realised that our convention of a `Base...` service would not be strictly applicable here, since the routing `ViewInvoiceService` would not be extending it, breaking the convention we use elsewhere. So we instead go with a convention of having a `Common` service, which the sroc/presroc services extend.
